requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==23.0.0
